Move VerseSelector component.
